requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0

# Data Analysis & AI/ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0

# Technical Analysis
yfinance>=0.2.28
alpha-vantage>=2.3.1
ta>=0.10.2

# Data Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Environment & Config
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.0

# JSON handling
pydantic>=2.5.0

# Async support
aiohttp>=3.9.0
